Edubot/requirements.txt
streamlit>=1.28.0
transformers>=4.30.0
torch>=2.0.0
scikit-learn>=1.3.0
pandas>=2.0.0
numpy>=1.24.0
requests>=2.31.0
beautifulsoup4>=4.12.0
python-dotenv>=1.0.0
youtube-search-python>=1.6.6
openai>=1.0.0
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.0
reportlab>=4.0.0
bcrypt>=4.0.0
email-validator>=2.0.0
pyjwt>=2.8.0
streamlit-authenticator>=0.2.3


